@@ -1,7 +1,8 @@
 /// Vertical alignment of content inside [SlideDrawerContainer]
 enum FlipDrawerAlignment {
-  /// content will be aligned vertically from the top 
-  start, 
+  /// content will be aligned vertically from the top
+  start,
+
   /// content will be centered vertically
-  center, 
+  center,
 }
